feat(compass-saved-aggregations-queries): clean up types
