+++ packages/compass-saved-aggregations-queries/src/stores/aggregations-queries-items.ts
@@ -1,7 +1,9 @@
 import type { Dispatch, Reducer } from 'redux';
 import toNS from 'mongodb-ns';
 import { FavoriteQueryStorage } from '@mongodb-js/compass-query-history';
+import type { Query } from '@mongodb-js/compass-query-history';
 import { PipelineStorage } from '@mongodb-js/compass-aggregations';
+import type { Aggregation } from '@mongodb-js/compass-aggregations';
 import type { ThunkAction } from 'redux-thunk';
 import type { RootState } from '.';
 
@@ -36,56 +38,6 @@ export type Item = {
       aggregation: Omit<Aggregation, 'lastModified'>;
     }
 );
-
-interface Query {
-  _id: string;
-  _name: string;
-  _ns: string;
-  _dateSaved: number;
-
-  collation?: Record<string, unknown>;
-  filter?: Record<string, unknown>;
-  limit?: number;
-  project?: Record<string, unknown>;
-  skip?: number;
-  sort?: Record<string, number>;
-}
-
-interface Aggregation {
-  id: string;
-  name: string;
-  namespace: string;
-  lastModified: number;
-
-  autoPreview?: boolean;
-  collation?: string;
-  collationString?: string;
-  comments?: boolean;
-  env?: string;
-  isReadonly?: boolean;
-  isTimeSeries?: boolean;
-  pipeline: Pipeline[];
-  sample?: boolean;
-  sourceName?: string;
-}
-
-interface Pipeline {
-  id: string;
-  stageOperator: string;
-  stage: string;
-  isValid: boolean;
-  isEnabled: boolean;
-  isExpanded: boolean;
-  isLoading: boolean;
-  isComplete: boolean;
-  previewDocuments: unknown[];
-  syntaxError: unknown;
-  error: unknown;
-  projections?: unknown[];
-  fromStageOperators?: boolean;
-  executor?: unknown;
-  isMissingAtlasOnlyStageSupport?: boolean;
-}
 
 export type State = {
   loading: boolean;
@@ -159,7 +111,7 @@ export const deleteItem = (
 };
 
 const getAggregationItems = async (): Promise<Item[]> => {
-  const aggregations: Aggregation[] = await pipelineStorage.loadAll();
+  const aggregations = await pipelineStorage.loadAll();
   return aggregations.map((aggregation) => {
     const { database, collection } = toNS(aggregation.namespace);
     return {
@@ -175,7 +127,7 @@ const getAggregationItems = async (): Promise<Item[]> => {
 };
 
 const getQueryItems = async (): Promise<Item[]> => {
-  const queries: Query[] = await favoriteQueryStorage.loadAll();
+  const queries = await favoriteQueryStorage.loadAll();
   return queries.map((query) => {
     const { database, collection } = toNS(query._ns);
     return {
